src/main/java/com/liwenwen/sell/service/Impl/ProductServiceImpl.java
package com.liwenwen.sell.service.Impl;

import com.liwenwen.sell.dao.ProductInfoDao;
import com.liwenwen.sell.enums.ProductStatusEnum;
import com.liwenwen.sell.pojo.ProductInfo;
import com.liwenwen.sell.service.ProductService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;


import java.util.List;
@Service
public class ProductServiceImpl implements ProductService {
    @Autowired
    ProductInfoDao productInfoDao;

    @Override
    public ProductInfo findOne(String productId) {
        return productInfoDao.getOne(productId);
    }

    @Override
    public Page<ProductInfo> findAll(Pageable pageable) {
        return productInfoDao.findAll((org.springframework.data.domain.Pageable) pageable);
    }

    @Override
    public List<ProductInfo> findUpAll() {
        return productInfoDao.findByProductStatus(ProductStatusEnum.UP.getCode());
    }

    @Override
    public ProductInfo save(ProductInfo productInfo) {
        return productInfoDao.save(productInfo);
    }
}
